refactor: fix all the problems arises from migrating from old services to sending commands and queries in the controllers

+++ src/SocialMedia.Application/Comments/Commands/Delete/DeleteCommentCommandHandler.cs
@@ -28,12 +28,21 @@ public class DeleteCommentCommandHandler(IUserService userService, IUnitOfWork u
         if (comment.UserId != user.Id)
             return Result.Failure(CommentErrors.Unauthorized);
 
-        if (comment.Replies is not null)
+        if (comment.Replies is not null && comment.RepliesCount > 0)
             unitOfWork.Comments.RemoveRange(comment.Replies);
+        else
+        {
+            // decrease the replies count of the parent comment it exists because we don't have multiple levels of comments
+            var parentComment = await unitOfWork.Comments.GetAsync(c => c.Id == comment.ParentCommentId);
+            if (parentComment != null)
+            {
+                parentComment.RepliesCount--;
+            }
+        }
 
         unitOfWork.Comments.Remove(comment);
         post.CommentsCount -= 1 + comment.RepliesCount;
-        comment.RaiseDomainEvent(() => new CommentDeletedDomainEvent(comment.Id, comment.PostId, comment.RepliesCount));
+        post.RaiseDomainEvent(() => new CommentDeletedDomainEvent(comment.Id, comment.PostId, comment.RepliesCount));
 
         var entitiesDeleted = await unitOfWork.SaveChangesAsync();
         if (entitiesDeleted == 0)
